+++ src/modules/CRUD/services/CRUD.js
@@ -49,8 +49,22 @@ export const getLayer = ({ menu = [] }, name) => {
     form_schema: schema,
     ui_schema: uiSchema,
     templates,
+    extent,
+    feature_list_properties: featureListProperties,
   } = flattenMenu(menu).find(item => item.layer.name === name) || {};
-  return layer ? { ...layer, settings, displayName, schema, uiSchema, templates } : false;
+  if (!layer) {
+    return false;
+  }
+  return {
+    ...layer,
+    settings,
+    displayName,
+    schema,
+    uiSchema,
+    templates,
+    extent,
+    featureListProperties,
+  };
 };
 
 export const getSources = ({ menu = [] }) =>